Reformat Urho3D source files to get rid of the indents on empty lines. Reorganize includes. Misc. fixes to perform explicit C-style cast on type mismatch. Misc. fixes on bugs identified by CLion's static analyzer.

+++ Source/Urho3D/Container/Hash.h
@@ -30,13 +30,13 @@ namespace Urho3D
 /// Pointer hash function.
 template <class T> unsigned MakeHash(T* value)
 {
-    return ((unsigned)(size_t)value) / sizeof(T);
+    return (unsigned)((size_t)value / sizeof(T));
 }
 
 /// Const pointer hash function.
 template <class T> unsigned MakeHash(const T* value)
 {
-    return ((unsigned)(size_t)value) / sizeof(T);
+    return (unsigned)((size_t)value / sizeof(T));
 }
 
 /// Generic hash function.
@@ -46,61 +46,61 @@ template <class T> unsigned MakeHash(const T& value)
 }
 
 /// Void pointer hash function.
-template<> inline unsigned MakeHash(void* value)
+template <> inline unsigned MakeHash(void* value)
 {
     return (unsigned)(size_t)value;
 }
 
 /// Const void pointer hash function.
-template<> inline unsigned MakeHash(const void* value)
+template <> inline unsigned MakeHash(const void* value)
 {
     return (unsigned)(size_t)value;
 }
 
 /// Long long hash function.
-template<> inline unsigned MakeHash(const long long& value)
+template <> inline unsigned MakeHash(const long long& value)
 {
-    return (value >> 32) | (value & 0xffffffff);
+    return (unsigned)((value >> 32) | (value & 0xffffffff));
 }
 
 /// Unsigned long long hash function.
-template<> inline unsigned MakeHash(const unsigned long long& value)
+template <> inline unsigned MakeHash(const unsigned long long& value)
 {
-    return (value >> 32) | (value & 0xffffffff);
+    return (unsigned)((value >> 32) | (value & 0xffffffff));
 }
 
 /// Int hash function.
-template<> inline unsigned MakeHash(const int& value)
+template <> inline unsigned MakeHash(const int& value)
 {
-    return value;
+    return (unsigned)value;
 }
 
 /// Unsigned hash function.
-template<> inline unsigned MakeHash(const unsigned& value)
+template <> inline unsigned MakeHash(const unsigned& value)
 {
     return value;
 }
 
 /// Short hash function.
-template<> inline unsigned MakeHash(const short& value)
+template <> inline unsigned MakeHash(const short& value)
 {
-    return value;
+    return (unsigned)value;
 }
 
 /// Unsigned short hash function.
-template<> inline unsigned MakeHash(const unsigned short& value)
+template <> inline unsigned MakeHash(const unsigned short& value)
 {
     return value;
 }
 
 /// Char hash function.
-template<> inline unsigned MakeHash(const char& value)
+template <> inline unsigned MakeHash(const char& value)
 {
-    return value;
+    return (unsigned)value;
 }
 
 /// Unsigned char hash function.
-template<> inline unsigned MakeHash(const unsigned char& value)
+template <> inline unsigned MakeHash(const unsigned char& value)
 {
     return value;
 }
